renamed packages to naming conventions
