+++ library-backend/src/main/java/com/lbdg/library_backend/repositories/BookRepository.java
@@ -1,6 +1,6 @@
-package com.lbdg.library_backend.Repositories;
+package com.lbdg.library_backend.repositories;
 
-import com.lbdg.library_backend.Entities.BookEntity;
+import com.lbdg.library_backend.entities.BookEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRepository extends JpaRepository<BookEntity, Long> {
